% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inspection.R
\name{do_check}
\alias{do_check}
\title{Check range}
\usage{
do_check(
  df,
  vars,
  check,
  label_check,
  template_ok = "all variables are {label_check}",
  template_fail = "{var} is not {label_check}"
)
}
\arguments{
\item{df}{Data}

\item{vars}{Character vector of variables to check}

\item{min}{Minimum allowed value (default: 0)}

\item{max}{Maximum allowed value (default: 1)}
}
\value{
List containing the results of the check (checks), and a tibble
of status and message for each test (messages).
}
\description{
Checks whether variables always fall within a given range.
}
\details{
Note that both the minimum and maximum are _inclusive_, that is,
the range is given as [min, max]. For _exclusive_ checks, e.g. (0,1], or
greater than 0, lesser than or equal to 1, the user will have to manually
give a minimum value with a certain error applied (e.g., 1e-6).

The list of messages in the result contains a single line if the test passed,
or if a test failed for one or more variables, a line for each failure.
}
\examples{
data(df_pa)
check_range(df_pa, c("u_pfs", "p_pfspd"))
}
