% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_metamodel.R
\name{dsa_lm_metamodel}
\alias{dsa_lm_metamodel}
\title{Perform DSA using linear metamodel}
\usage{
dsa_lm_metamodel(df, lm_metamodel)
}
\arguments{
\item{df}{a dataframe. This dataframe should contain both the dependent and independent variables of `lm_metamodel`.}

\item{lm_metamodel}{a lm object. This object should use variables defined in `df`.}
}
\value{
A dataframe with the results of deterministic sensitivity analyses performed using parameter values of the linear metamodel. The dataframe contains the results using the lower and upper bound of the 95% Confidence Interval of the probabilistic parameters.
}
\description{
This function performs deterministic sensitivity analyses (DSA) using the results of a linear metamodel. (STILL IN DEVELOPMENT)
}
\details{
The details of the methods used in for these DSA are described in XXX.
}
\examples{
# Fitting meta model with two variables using the summary data
data(df_pa)
lm_res_2 <- fit_lm_metamodel(df = df_pa,
                 y = "Inc_QALY",
                 x = c("p_pfsd", "p_pdd")
                 )

dsa_lm_metamodel(df = df_pa,
                 lm_metamodel = lm_res_2)

}
