% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hemodel.R
\name{generate_det_inputs}
\alias{generate_det_inputs}
\title{Generate deterministic model inputs.}
\usage{
generate_det_inputs()
}
\value{
A list. A description of the inputs parameters is available in the documentation of the \code{\link{df_pa}} dataframe.
}
\description{
This function generates the deterministic model inputs for the example health economic model developed to test the functionalities of the package.
}
\examples{
# Generating deterministic model inputs and storing them in an object.
l_inputs_det <- generate_det_inputs()
}
