% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_l_iviRA_pa_params.R
\docType{data}
\name{l_iviRA_pa_params}
\alias{l_iviRA_pa_params}
\title{Input list for testing from iviRA R package}
\format{
A list containing all necessary inputs of the iviRA health economic model.
}
\source{
Incerti D, Curtis JR, Shafrin J, Lakdawalla DN, Jansen JP (2019). “ A Flexible Open-Source Decision Model for Value Assessment of Biologic Treatment for Rheumatoid Arthritis.” PharmacoEconomics. doi: 10.1007/s40273-018-00765-2.

https://innovationvalueinitiative.github.io/IVI-RA/index.html
}
\usage{
l_iviRA_pa_params
}
\description{
See the documentation of the \code{\link[iviRA:sample_pars]{iviRA::sample_pars}} function for details.
}
\details{
A list containing 1,000 sets of probabilistic inputs of a individual-level health state transition model developed using \code{\link[iviRA]{iviRA}} package.
}
\keyword{datasets}
