% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_l_psa_aaa.R
\docType{data}
\name{l_psa_aaa}
\alias{l_psa_aaa}
\title{A dataframe containing probabilistic inputs for testing}
\format{
The dataframe contains the 1,000 probabilistic inputs and outputs obtained by executing the Main file_men 30years_FullModel.R' script (for 1,000 individuals) under the 'models' folder on the github repository of the model.
}
\source{
Sweeting MJ, Masconi KL, Jones E, Ulug P, Glover MJ, Michaels JA, Bown MJ, Powell JT, Thompson SG. Analysis of clinical benefit, harms, and cost-effectiveness of screening women for abdominal aortic aneurysm. Lancet. 2018 Aug 11;392(10146):487-495. doi: 10.1016/S0140-6736(18)31222-4.

Link to Github repository: \url{https://github.com/mikesweeting/AAA_DES_model}
}
\usage{
l_psa_aaa
}
\description{
A list containing 1,000 sets of probabilistic inputs and outputs from a discrete event simulation developed in R.
}
\keyword{datasets}
