% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_metamodel.R
\name{predict_lm_metamodel}
\alias{predict_lm_metamodel}
\title{Predict using linear metamodel}
\usage{
predict_lm_metamodel(lm_metamodel, inputs)
}
\arguments{
\item{lm_metamodel}{a lm object. This object should use variables defined in `df`.}

\item{inputs}{a numeric value or vector of numeric values. These inputs value will be used for the prediction using the metamodel.}
}
\value{
A dataframe with the results of deterministic sensitivity analyses performed using parameter values of the linear metamodel. The dataframe contains the results using the lower and upper bound of the 95% Confidence Interval of the probabilistic parameters.
}
\description{
This function computes a result using a pre-defined linear metamodel, and user-defined inputs to make the prediction.
}
\details{
The number of element of `inputs` should equal the number of predictors included in `lm_metamodel`.
}
\examples{
# Fitting meta modelwith two variables using the summary data
data(df_pa)
lm_res <- fit_lm_metamodel(df = df_pa,
                 y = "Inc_QALY",
                 x = c("p_pfsd", "p_pdd")
                 )

# Predicting using this metamodel
predict_lm_metamodel(lm_metamodel = lm_res,
                     inputs = c(0.75, 0.2)
                     )

}
