% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_costeffectiveness.R
\name{summary_ice}
\alias{summary_ice}
\title{Summary statistics of the incremental cost-effectiveness plane.}
\usage{
summary_ice(df, inc_e, inc_c)
}
\arguments{
\item{df}{a dataframe.}

\item{inc_e}{character string. Name of variable containing the incremental effects.}

\item{inc_c}{character string. Name of variable containing the incremental costs}
}
\value{
A dataframe.
}
\description{
This function computes the probability that the probabilistic outcome is in each of the quadrant.
}
\examples{
# Generating statistics of the incremental cost-effectiveness plane using the example data.
data(df_pa)
plot_ice(df = df_pa,
         inc_e = "inc_qaly",
         inc_c = "inc_costs")

}
