% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_costeffectiveness.R
\name{calculate_nb}
\alias{calculate_nb}
\title{Calculate NMB and NHB.}
\usage{
calculate_nb(df, e_int, e_comp, c_int, c_comp, wtp)
}
\arguments{
\item{df}{a dataframe.}

\item{e_int}{character. Name of variable of the dataframe containing total effects of the intervention strategy.}

\item{e_comp}{character. Name of variable of the dataframe containing total effects of the comparator strategy.}

\item{c_int}{character. Name of variable of the dataframe containing total costs of the intervention strategy.}

\item{c_comp}{character. Name of variable of the dataframe containing total costs of the comparator strategy.}

\item{wtp}{numerical values. Willingness-to-pay thresholds to use for NMB and NHB calculations.}
}
\value{
A dataframe.
}
\description{
This function calculates the Net Monetary Benefits (NMB) and Net Health Benefits (NHB) for each strategy and the incremental NMB and NHB.
}
\examples{
# Calculate NB's at a willingness-to-pay threshold of 80000 per unit of effects
data("df_pa")
calculate_nb(df_pa,
             e_int = "QALY_int",
             e_comp = "QALY_comp",
             c_int = "Costs_int",
             c_comp = "Costs_comp",
             wtp = 80000)

}
