% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_df_ckd_inputs.R
\docType{data}
\name{df_ckd_inputs}
\alias{df_ckd_inputs}
\title{A dataframe containing probabilistic inputs for testing}
\format{
The dataframe contains the 1,000 probabilistic inputs (indexes 1 to 1000) from the PSA.xlsx file available on the github repository of the model.
}
\source{
Marika M. Cusick, Rebecca L. Tisdale, Glenn M. Chertow, et al. Population-Wide Screening for Chronic Kidney Disease: A Cost-Effectiveness Analysis. Ann Intern Med.2023;176:788-797. [Epub 23 May 2023]. doi:10.7326/M22-3228.

Link to Github repository: \url{https://github.com/marikamaecusick/CKDScreeningCEA}
}
\usage{
df_ckd_inputs
}
\description{
A dataframe containing 1,000 sets of probabilistic inputs from a health state transition model developed in Python.
}
\keyword{datasets}
