% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_df_ckd_results.R
\docType{data}
\name{df_ckd_results}
\alias{df_ckd_results}
\title{A dataframe containing probabilistic outputs for testing}
\format{
The dataframe contains the 1,000 probabilistic outputs from the CKD Screening CEA health economic model.
}
\source{
Marika M. Cusick, Rebecca L. Tisdale, Glenn M. Chertow, et al. Population-Wide Screening for Chronic Kidney Disease: A Cost-Effectiveness Analysis. Ann Intern Med.2023;176:788-797. [Epub 23 May 2023]. doi:10.7326/M22-3228.

https://github.com/marikamaecusick/CKDScreeningCEA
}
\usage{
df_ckd_results
}
\description{
A dataframe containing 1,000 sets of probabilistic outputs from a health state transition model developed in Python for individuals aged 35, screened every 10 years, versus no screening.
}
\keyword{datasets}
