% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_iviRA_pa.R
\docType{data}
\name{df_iviRA_pa}
\alias{df_iviRA_pa}
\title{Dataframe of inputs and outputs of a health economic model developed and evaluated with the iviRA R package for testing}
\format{
See the documentation of the `sample_pars` function of the iviRA package for details.
}
\source{
Incerti D, Curtis JR, Shafrin J, Lakdawalla DN, Jansen JP (2019). “ A Flexible Open-Source Decision Model for Value Assessment of Biologic Treatment for Rheumatoid Arthritis.” PharmacoEconomics. doi: 10.1007/s40273-018-00765-2.

Link to Github repository: \url{https://innovationvalueinitiative.github.io/IVI-RA/index.html}

Link to iviRA website: \url{https://innovationvalueinitiative.github.io/IVI-RA/}
}
\usage{
df_iviRA_pa
}
\description{
A dataframe containing 1,000 sets of probabilistic inputs and outputs of a individual-level health state transition model developed using iviRA package available on Github.
}
\keyword{datasets}
