% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_df_pa.R
\docType{data}
\name{df_pa}
\alias{df_pa}
\title{Dataframe for testing}
\format{
An object of class \code{data.frame} with 10000 rows and 18 columns.
}
\usage{
df_pa
}
\description{
A dataframe containing 10,000 iterations (inputs and outputs) of a probabilistic analysis
}
\keyword{datasets}
