% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lasso.R
\name{fit_lasso_metamodel}
\alias{fit_lasso_metamodel}
\title{Fit LASSO metamodel}
\usage{
fit_lasso_metamodel(
  df,
  y_var = NULL,
  x_vars = NULL,
  seed_num = 1,
  standardise = FALSE,
  tune_plot = TRUE,
  x_poly_2 = NULL,
  x_poly_3 = NULL,
  x_exp = NULL,
  x_log = NULL,
  x_inter = NULL
)
}
\arguments{
\item{fit_complete_model}{}
}
\description{
Fit LASSO metamodel
}
