% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_metamodel.R
\name{fit_lm_metamodel}
\alias{fit_lm_metamodel}
\title{Fit linear metamodel}
\usage{
fit_lm_metamodel(df, y, x)
}
\arguments{
\item{df}{a dataframe.}

\item{y}{character. Name of the output variable in the dataframe. This will be the dependent variable of the metamodel.}

\item{x}{character or a vector for characters. Name of the input variable in the dataframe. This will be the independent variable of the metamodel.}
}
\value{
A dataframe with summary data for selected inputs and outputs.
}
\description{
This function fits and provides summary statistics of a linear regression model fitted on the input and output values of a probabilistic analysis.
}
\examples{
#' # Fitting meta model with a single variable using the summary data
data(df_pa)
fit_lm_metamodel(df = df_pa,
                 y = "Inc_QALY",
                 x = "p_pfsd")
                 )

# Fitting meta model with two variables using the summary data
data(df_pa)
fit_lm_metamodel(df = df_pa,
                 y = "Inc_QALY",
                 x = c("p_pfsd", "p_pdd")
                 )

}
