% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_randomforest.R
\name{fit_rf_metamodel}
\alias{fit_rf_metamodel}
\title{Fit random forest metamodel}
\usage{
fit_rf_metamodel(
  df,
  y_var = NULL,
  x_vars = NULL,
  standardise = FALSE,
  x_poly_2 = NULL,
  x_poly_3 = NULL,
  x_exp = NULL,
  x_log = NULL,
  x_inter = NULL
)
}
\arguments{
\item{x_inter}{}
}
\description{
Fit random forest metamodel
}
