% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inspection.R
\name{generate_cor}
\alias{generate_cor}
\title{Generate correlation matrix}
\usage{
generate_cor(df, v_params = "ALL")
}
\arguments{
\item{df}{a dataframe.}

\item{v_params}{character or vector of character. Vector of names of the inputs and outputs for which to return the correlation matrix. Default is "ALL" which returns the correlation matrix for all inputs and outputs in the dataframe.}
}
\value{
A table with summary data for selected inputs and outputs.
}
\description{
This function generates the correlation matrix of input and output values of a probabilistic analysis.
}
\examples{
# Generating summary data of all inputs using the example dataframe
data(df_pa)
generate_cor(df_pa)

}
