% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hemodel.R
\name{generate_pa_inputs_psm}
\alias{generate_pa_inputs_psm}
\title{Generate probabilistic model inputs for partitioned survival model.}
\usage{
generate_pa_inputs_psm(n_sim = 10000, sd_var = 0.2, seed_num = 452)
}
\value{
A dataframe. A description of the variables of the returned dataframe is available in the documentation of the \code{\link{df_pa_psm}} dataframe.
}
\description{
This function generates the probabilistic model inputs for the example health economic model developed to test the functionalities of the package.
}
\examples{
# Generating deterministic model inputs and storing them in an object.
df_inputs_prob <- generate_pa_inputs_psm()
}
