% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inspection.R
\name{generate_sum_stats}
\alias{generate_sum_stats}
\title{Generate summary statistics}
\usage{
generate_sum_stats(df, v_params = "ALL")
}
\arguments{
\item{df}{a dataframe.}

\item{v_params}{character or vector of character. Vector of names of the variables of `df` for which to return summary statistics. Default is "ALL" which returns summary values for all inputs and outputs in the dataframe.}
}
\value{
A dataframe with summary data for the selected variables.
}
\description{
This function generates summary statistics of input and output values of a probabilistic analysis.
}
\examples{
# Generating summary data of all inputs using the example dataframe
data(df_pa)
df_summary <- generate_sum_stats(df_pa)

}
