% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_l_iviRA_out_summ_1.R
\docType{data}
\name{l_iviRA_out_summ_1}
\alias{l_iviRA_out_summ_1}
\title{List of summary outputs  for testing from iviRA R package
A list containing 1,000 sets of probabilistic outputs of a individual-level health state transition model developed using \code{\link[iviRA]{iviRA}} package.}
\format{
A list containing the summary outputs a treatment sequence ("adamtx", "etnmtx", "ifxmtx") of the iviRA health economic model.
}
\source{
Incerti D, Curtis JR, Shafrin J, Lakdawalla DN, Jansen JP (2019). “ A Flexible Open-Source Decision Model for Value Assessment of Biologic Treatment for Rheumatoid Arthritis.” PharmacoEconomics. doi: 10.1007/s40273-018-00765-2.

https://innovationvalueinitiative.github.io/IVI-RA/index.html
}
\usage{
l_iviRA_out_summ_1
}
\description{
See the documentation of the \code{\link[iviRA:sim_iviRA]{iviRA::sim_iviRA}} function for details.
}
\keyword{datasets}
