% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_costeffectiveness.R
\name{plot_ceac}
\alias{plot_ceac}
\title{Plot cost-effectiveness acceptability curves.}
\usage{
plot_ceac(df, wtp)
}
\arguments{
\item{df}{a dataframe.}

\item{wtp}{vector of numerical values. Name of variable of the dataframe containing the willingness-to-pay thresholds at which the probability of cost effectiveness have been defined.}
}
\value{
A ggplot graph.
}
\description{
This function calculates the probabilities that each strategy is the most cost effective.
}
\examples{
# Plot CEAC based on results from calculate_ceac()
data("df_pa")
m_res_ceac <- calculate_ceac(df_pa,
              e_int = "QALY_int",
              e_comp = "QALY_comp",
              c_int = "Costs_int",
              c_comp = "Costs_comp")
df_ceac_p <- as.data.frame(m_res_ceac)

plot_ceac(df = df_ceac_p,
          v_wtp = "WTP_threshold")

}
