% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inspection.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot moving average}
\usage{
plot_convergence(df, outcome, block_size = 500, conv_limit = 0.01)
}
\arguments{
\item{df}{a dataframe.}

\item{outcome}{character string. Name of variable of the dataframe for which to plot the moving average.}

\item{block_size}{numeric. Define the number of iterations at which the mean outcome has to be defined and plotted. NOT USED YET!}

\item{conv_limit}{numeric. Define the convergence limit, under which the relative change between block of iterations should lie.  NOT USED YET!}
}
\value{
A ggplot graph.
}
\description{
This function plots the moving average of a outcome variable
}
\examples{
# Checking the moving average of the incremental QALYs using the example data.
data(df_pa)
plot_convergence(df = df_pa,
                 outcome = "Inc_QALY"
                 ))

}
