% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inspection.R
\name{plot_ice}
\alias{plot_ice}
\title{Plotting the incremental cost-effectiveness plane.}
\usage{
plot_ice(df, param_1, param_2, wtp = NULL)
}
\arguments{
\item{df}{a dataframe.}

\item{param_1}{character. Name of variable of the dataframe to be plotted on the x-axis.}

\item{param_2}{character. Name of variable of the dataframe to be plotted on the y-axis.}

\item{wtp}{numeric. Default is NULL. If different than NULL, plots a linear line with intercept 0 and the defined slope.}
}
\value{
A ggplot graph.
}
\description{
This function plots the incremental cost-effectiveness plot.
}
\examples{
Generating plot using the example dataframe, and a willlingness-to-pay threshold of 80,0000 euros.
data(df_pa)
plot_ice(df = df_pa, param_1 = "Inc_QALY", param_1 = "Inc_Costs")

}
