% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inspection.R
\name{vis_1_param}
\alias{vis_1_param}
\title{Visualise the distribution of a single parameter}
\usage{
vis_1_param(df, param, binwidth = NULL, type = "histogram")
}
\arguments{
\item{df}{a dataframe.}

\item{param}{character. Name of variable of the dataframe for which the distribution should be plotted.}

\item{binwidth}{numeric. Determine the width of the bins to use, only applied in combination with "histogram". Default is 30 bins.}

\item{type}{character. Determine which plot to return: "histogram" for a histogram, "density" for a density plot. Default is "histogram".}
}
\value{
A ggplot graph.
}
\description{
This function plots the distribution of a single parameter.
}
\examples{
# Generating plot using the example dataframe for the costs of progression-free health state, and bins of 50 euros.
data(df_pa)
vis_1_param(df = df_pa, param = "c_pfs", binwidth = 50)

}
