% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_inspection.R
\name{vis_2_params}
\alias{vis_2_params}
\title{Visualise the distribution of two parameters}
\usage{
vis_2_params(df, param_1, param_2, slope = NULL, check = NULL)
}
\arguments{
\item{df}{a dataframe.}

\item{param_1}{character. Name of variable of the dataframe to be plotted on the x-axis.}

\item{param_2}{character. Name of variable of the dataframe to be plotted on the y-axis.}

\item{slope}{numeric. Default is NULL. If different than 0, plots a linear line with intercept 0 and the defined slope.}

\item{check}{character. Default is NULL. When set to "param_2 > param_1", plots dot fulfilling the condition in red.}
}
\value{
A ggplot graph.
}
\description{
This function plots the distribution of two parameters in a scatterplot.
}
\examples{
# Generating plot using the example dataframe for the costs of progression-free health state, and bins of 50 euros.
data(df_pa)
vis_1_param(df = df_pa, param = "c_pfs", binwidth = 50)

}
